+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.42.0] - 2025-04-28
+
 ### Forge
 
 #### Added
@@ -25,9 +27,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Cast
 
+#### Changed
+
+- when using `sncast call` the response will be printed as a Cairo-like string representation of the return values
+
 #### Fixed
 
 - Bug that prevented from passing values to `--arguments` that started with a leading minus `-` sign.
+
+### Cast
+
+### Fixed
+
+- User is now prompted to save an imported or deployed account in `sncast` config even when using `--network` flag
 
 ## [0.41.0] - 2025-04-08